chore: consolidate two methods into one
